Add statistics assuming unique orders by users
This is needed as a reader can take the same book not only once,
so in orders there can be duplicate orders having the same book
and the same reader and being differrent only by the time stamp.
So, our statistics methods should vary to let us know top books
by orders and by readers respectively.