Refactor library class statistics methods
Now all 5 interface methods use a single private method which groups
orders by any field, counts the number of orders by that entity and
returns the specified number of those counts.

Also it takes an optional block for the possibility of filtering the
initial orders' collections, e.g. only taking unique combinations of
a book and a reader.